r/aws_batch_job_definition: handle `INACTIVE` definitions in `List`
Adds handling to skip results which the `Read` operation will consider `NotFound`. Also moves list handling to it's own source file and adds more explicit checks for the identity returned by `List`.

```console
% make t K=batch T=TestAccBatchJobDefinition_List
make: Verifying source code with gofmt...
==> Checking that code complies with gofmt requirements...
make: Running acceptance tests on branch: 🌿 b-batch_job_definition-list 🌿...
TF_ACC=1 go1.25.5 test ./internal/service/batch/... -v -count 1 -parallel 20 -run='TestAccBatchJobDefinition_List'  -timeout 360m -vet=off
2025/12/22 14:23:05 Creating Terraform AWS Provider (SDKv2-style)...
2025/12/22 14:23:05 Initializing Terraform AWS Provider (SDKv2-style)...
=== RUN   TestAccBatchJobDefinition_List_basic
=== PAUSE TestAccBatchJobDefinition_List_basic
=== CONT  TestAccBatchJobDefinition_List_basic
--- PASS: TestAccBatchJobDefinition_List_basic (30.60s)
PASS
ok      github.com/hashicorp/terraform-provider-aws/internal/service/batch      36.959s
```
