+++ .changelog/45694.txt
@@ -1,0 +1,3 @@
+```release-note:bug
+resource/aws_batch_job_definition: Fix handling of logically deleted results in List
+```
