+++ internal/service/batch/job_definition_list_test.go
@@ -16,6 +16,7 @@ import (
 	"github.com/hashicorp/terraform-plugin-testing/tfversion"
 	"github.com/hashicorp/terraform-provider-aws/internal/acctest"
 	tfknownvalue "github.com/hashicorp/terraform-provider-aws/internal/acctest/knownvalue"
+	tfstatecheck "github.com/hashicorp/terraform-provider-aws/internal/acctest/statecheck"
 	"github.com/hashicorp/terraform-provider-aws/names"
 )
 
@@ -25,6 +26,9 @@ func TestAccBatchJobDefinition_List_basic(t *testing.T) {
 	resourceName1 := "aws_batch_job_definition.test[0]"
 	resourceName2 := "aws_batch_job_definition.test[1]"
 	rName := sdkacctest.RandomWithPrefix(acctest.ResourcePrefix)
+
+	arn1 := tfstatecheck.StateValue()
+	arn2 := tfstatecheck.StateValue()
 
 	acctest.ParallelTest(ctx, t, resource.TestCase{
 		TerraformVersionChecks: []tfversion.TerraformVersionCheck{
@@ -42,7 +46,10 @@ func TestAccBatchJobDefinition_List_basic(t *testing.T) {
 					acctest.CtRName: config.StringVariable(rName),
 				},
 				ConfigStateChecks: []statecheck.StateCheck{
+					arn1.GetStateValue(resourceName1, tfjsonpath.New(names.AttrARN)),
 					statecheck.ExpectKnownValue(resourceName1, tfjsonpath.New(names.AttrARN), tfknownvalue.RegionalARNExact("batch", "job-definition/"+rName+"-0:1")),
+
+					arn2.GetStateValue(resourceName2, tfjsonpath.New(names.AttrARN)),
 					statecheck.ExpectKnownValue(resourceName2, tfjsonpath.New(names.AttrARN), tfknownvalue.RegionalARNExact("batch", "job-definition/"+rName+"-1:1")),
 				},
 			},
@@ -57,11 +64,11 @@ func TestAccBatchJobDefinition_List_basic(t *testing.T) {
 				},
 				QueryResultChecks: []querycheck.QueryResultCheck{
 					querycheck.ExpectIdentity("aws_batch_job_definition.test", map[string]knownvalue.Check{
-						names.AttrARN: knownvalue.NotNull(),
+						names.AttrARN: arn1.Value(),
 					}),
 
 					querycheck.ExpectIdentity("aws_batch_job_definition.test", map[string]knownvalue.Check{
-						names.AttrARN: knownvalue.NotNull(),
+						names.AttrARN: arn2.Value(),
 					}),
 				},
 			},
